% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cohens_f.R
\name{cohens_f}
\alias{cohens_f}
\title{Calculate Cohen's f.}
\usage{
cohens_f(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data set}
}
\value{
numerica value
}
\description{
Calculate Cohen's f.
}
\examples{
"no test yet"
}
