% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df_cancer}
\alias{df_cancer}
\title{Test data to run the examples}
\format{
A data frame with 2 variables:
\describe{
\item{treatment_group}{ }
\item{control_group}{ }
}
}
\usage{
df_cancer
}
\description{
A dataset that includes 120 individuals.
}
\keyword{datasets}
