% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df_income}
\alias{df_income}
\title{Test data to run the examples}
\format{
A data frame with 2 variables and 30 rows
\describe{
\item{monthly_income}{ }
\item{sex}{ }
}
}
\usage{
df_income
}
\description{
A dataset that includes 120 individuals with sex gender and monthly income.
}
\keyword{datasets}
