% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df_stress}
\alias{df_stress}
\title{Test data to run the examples}
\format{
A data frame with 2.
\describe{
\item{baseline_stress}{ }
\item{one_year_stress}{ }
}
}
\usage{
df_stress
}
\description{
A dataset that includes 120 individuals.
}
\keyword{datasets}
