% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample.R
\name{draw_sample}
\alias{draw_sample}
\title{Draw Samples}
\usage{
draw_sample(k_groups = 3, f = 0.25, sd = c(1, 1, 1), max_n = 50)
}
\arguments{
\item{k_groups}{number of groups (levels of factor_A)}

\item{f}{cohens f.}

\item{sd}{standard deviations of the goups}

\item{max_n}{sample size for the groups (total sample size = max_n*k_groups)}
}
\value{
returns a data.frame with the columns y (observations) and x (factor_A).
}
\description{
Draw exemplary samples for the sequential ANOVA function.
}
\examples{
set.seed(333)
data <- sprtt::draw_sample()
data

data <- sprtt::draw_sample(
  k_groups = 4,
  sd = c(1, 0.8, 1, 1),
  max_n = 30
)
data

}
