% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eta_squared.R
\name{eta_squared}
\alias{eta_squared}
\title{Calculate eta squared.}
\usage{
eta_squared(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data set}
}
\value{
numeric value
}
\description{
Calculate eta squared.
}
\examples{
"no test yet"
}
