% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_external.R
\name{get_sprtt_cache_dir}
\alias{get_sprtt_cache_dir}
\title{Get path to cached simulation data directory}
\usage{
get_sprtt_cache_dir()
}
\description{
Get path to cached simulation data directory
}
\keyword{internal}
