% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprt.R
\name{plot_sprt}
\alias{plot_sprt}
\title{Plot Results of SPRTs}
\usage{
plot_sprt(
  sample_size,
  lr_log,
  A_boundary_log,
  B_boundary_log,
  labels = TRUE,
  position_labels_x = 0.15,
  position_labels_y = 0.075,
  font_size = 15,
  line_size = 0.8
)
}
\arguments{
\item{sample_size}{sample size.}

\item{lr_log}{log-likelihood-ratio.}

\item{A_boundary_log}{Log of the A boundary.}

\item{B_boundary_log}{Log of the B boundary.}

\item{labels}{show labels.}

\item{position_labels_x}{position of the boundary labels on the x-axis}

\item{position_labels_y}{position of the boundary labels on the y-axis}

\item{font_size}{font size of the plot}

\item{line_size}{line size of the plot}
}
\value{
is a plot
}
\description{
Plot Results of SPRTs
}
\examples{
inst/examples/sprt_plot.R
}
