% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprtt.R
\docType{package}
\name{sprtt}
\alias{sprtt}
\title{sprtt: help page.}
\description{
This package provides the implementation of sequential probability
ratio tests using t-statistic.
}
\section{detailed help pages}{

For detailed instructions on the \code{sprtt} package, see:
\code{vignette("sprtt")} or https://meikesteinhilber.github.io/sprtt/"
}

\section{sprtt functions}{

\itemize{
\item \code{\link{seq_ttest}} Performs one and two sample sequential t-tests on vectors of
data.
\item \code{\link{seq_anova}} Performs a sequential one-way ANOVA on vectors of
data.
}
}

