% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_seq_anova_arguments.R
\name{[,seq_anova_arguments-method}
\alias{[,seq_anova_arguments-method}
\title{Method to retrieve the contents of a slot of an object of the
seq_anova_arguments class.}
\usage{
\S4method{[}{seq_anova_arguments}(x, i, j, drop)
}
\arguments{
\item{x}{the seq_anova_arguments object.}

\item{i}{indices indicating elements to extract.}

\item{j}{not used.}

\item{drop}{not used.}

\item{seq_anova_arguments}{the corresponding class to this method.}
}
\value{
Returns the contents of the specified slot. For more information,
see the documentation for the seq_anova_arguments class.
}
\description{
This method is only used internally to process
the input arguments of the \code{\link{seq_anova}} function. As a normal user,
you can ignore this specific documentation.
}
\keyword{internal}
