% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_seq_anova_results.R
\name{[,seq_anova_results,ANY,ANY,ANY-method}
\alias{[,seq_anova_results,ANY,ANY,ANY-method}
\title{Method to retrieve the contents of a slot of an object of the
\code{\linkS4class{seq_anova_results}} class.}
\usage{
\S4method{[}{seq_anova_results,ANY,ANY,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{the seq_ttest_results object.}

\item{i}{indices indicating elements to extract.}

\item{j}{not used.}

\item{drop}{not used.}

\item{seq_anova_results}{the corresponding class to this method.}
}
\value{
Returns the contents of the specified slot. For more information,
see the documentation for the \code{\linkS4class{seq_anova_results}} class.
}
\description{
Method to retrieve the contents of a slot of an object of the
\code{\linkS4class{seq_anova_results}} class.
}
\keyword{internal}
