% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_seq_ttest_arguments.R
\name{[,seq_ttest_arguments,ANY,ANY,ANY-method}
\alias{[,seq_ttest_arguments,ANY,ANY,ANY-method}
\title{Method to retrieve the contents of a slot of an object of the
seq_ttest_arguments class.}
\usage{
\S4method{[}{seq_ttest_arguments,ANY,ANY,ANY}(x, i, j, drop)
}
\arguments{
\item{x}{the seq_ttest_arguments object.}

\item{i}{indices indicating elements to extract.}

\item{j}{not used.}

\item{drop}{not used.}

\item{seq_ttest_arguments}{the corresponding class to this method.}
}
\value{
Returns the contents of the specified slot of an
seq_ttest_arguments object. For more information, see the arguments of the
\code{\link{seq_ttest}} function.
}
\description{
This method is only used internally to process
the input arguments of the \code{\link{seq_ttest}} function. As a normal user,
you can ignore this specific documentation.
}
\keyword{internal}
