% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_seq_ttest_arguments.R
\name{[,seq_ttest_arguments-method}
\alias{[,seq_ttest_arguments-method}
\title{Method to retrieve the contents of a slot of an object of the seq_ttest_arguments class.}
\usage{
\S4method{[}{seq_ttest_arguments}(x, i, j, drop)
}
\arguments{
\item{x}{the seq_ttest_arguments object.}

\item{i}{the start of the string of the slot name.}

\item{j}{the end of the string of the slot name.}

\item{drop}{not used.}

\item{seq_ttest_arguments}{the corresponding class to this method.}
}
\description{
Method to retrieve the contents of a slot of an object of the seq_ttest_arguments class.
}
