% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLASS_seq_ttest_results.R
\name{[,seq_ttest_results-method}
\alias{[,seq_ttest_results-method}
\title{Method to retrieve the contents of a slot of an object of the
seq_ttest_results class.}
\usage{
\S4method{[}{seq_ttest_results}(x, i, j, drop)
}
\arguments{
\item{x}{the seq_ttest_results object.}

\item{i}{the start of the string of the slot name.}

\item{j}{the end of the string of the slot name.}

\item{drop}{not used.}

\item{seq_ttest_results}{the corresponding class to this method.}
}
\value{
the content of the slot.
}
\description{
Method to retrieve the contents of a slot of an object of the
seq_ttest_results class.
}
