% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_external.R
\name{cache_clear}
\alias{cache_clear}
\title{Clear cached simulation data}
\usage{
cache_clear()
}
\description{
Removes locally cached simulation data. Data will be re-downloaded on next use.
}
\examples{
\dontrun{
# Clear cache
cache_clear()
}
}
