% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df_income}
\alias{df_income}
\title{Test data to run the examples}
\format{
A data frame with 5 variables and 120 rows
\describe{
\item{monthly_income}{ }
\item{sex}{ male or female}
}
}
\usage{
df_income
}
\description{
A dataset that includes 120 individuals with their gender, monthly income,
and information about children and debt.
}
\keyword{datasets}
