% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df_income}
\alias{df_income}
\title{Test data to run the examples}
\format{
A data frame with 5 variables and 30 rows
\describe{
\item{monthly_income}{ }
\item{in_debt}{0 = no and 1 = yes}
\item{sex}{ male or female}
\item{age}{ }
}
}
\usage{
df_income
}
\description{
A dataset that includes 30 individuals with their gender, monthly income,
and information about children and debt.
}
\keyword{datasets}
