% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\docType{data}
\name{df_stress}
\alias{df_stress}
\title{Test data to run the examples}
\format{
A data frame with 4 variables and 30 rows
\describe{
\item{treatment_group}{ }
\item{control_group}{ }
\item{baseline_stress}{ }
\item{one_year_stress}{ }
}
}
\usage{
df_stress
}
\description{
A dataset that includes 30 individuals.
}
\keyword{datasets}
