% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sample_mixture.R
\name{draw_sample_mixture}
\alias{draw_sample_mixture}
\title{Draw Samples from a Gaussian Mixture Distribution}
\usage{
draw_sample_mixture(k_groups, f, max_n, counter_n = 100, verbose = FALSE)
}
\arguments{
\item{k_groups}{number of groups (levels of factor_A)}

\item{f}{Cohen's f. The simulated effect size.}

\item{max_n}{sample size for the groups (total sample size = max_n*k_groups)}

\item{counter_n}{number of times the functions. Default value is 1 for each group.}

\item{sample_ratio}{sample ratio between th groups. Default value is 1 for each group.}
}
\value{
returns a data.frame with the columns y (observations) and x (factor_A).
}
\description{
Draws exemplary samples with a certain effect size for the sequential one-oway ANOVA or the sequential t-test.
}
\examples{
set.seed(333)

data <- sprtt::draw_sample_mixture(
  k_groups = 2,
  f = 0.40,
  max_n = 30
)
data

data <- sprtt::draw_sample_mixture(
  k_groups = 4,
  f = 0,
  max_n = 25,
  verbose = TRUE,
  counter_n = 150
)
data
}
