% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cohens_f.R
\name{effect_sizes}
\alias{effect_sizes}
\title{Calculate effect sizes.}
\usage{
effect_sizes(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data set}
}
\value{
numerica value
}
\description{
Calculate effect sizes.
}
\examples{
"no test yet"
}
