% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sprt.R
\name{plot_sprt}
\alias{plot_sprt}
\title{Plot Results of SPRTs}
\usage{
plot_sprt(
  sprt_results,
  labels = TRUE,
  position_labels_x = 0.15,
  position_labels_y = 0.075,
  font_size = 25,
  line_size = 1.5,
  highlight_color = "#CD2626"
)
}
\arguments{
\item{sprt_results}{result object of a seq_ttest() or a seq_anova()}

\item{labels}{show labels in the plot.}

\item{position_labels_x}{position of the boundary labels on the x-axis}

\item{position_labels_y}{position of the boundary labels on the y-axis}

\item{font_size}{font size of the plot}

\item{line_size}{line size of the plot}

\item{highlight_color}{highlighting color, default is red.}
}
\value{
returns a plot
}
\description{
Plot Results of SPRTs
}
\examples{
inst/examples/sprt_plot.R
}
