% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_anova.R
\name{seq_anova}
\alias{seq_anova}
\title{Sequential Analysis of Variance}
\usage{
seq_anova(formula, f, alpha = 0.05, power = 0.95, data, verbose = TRUE)
}
\arguments{
\item{formula}{a formula.}

\item{f}{cohens f.}

\item{alpha}{the type I error. A number between 0 and 1.}

\item{power}{1 - beta (beta is the type II error probability). A number
between 0 and 1.}

\item{data}{data frame.}

\item{verbose}{a logical value whether you want a verbose output or not.}
}
\value{
An object of the S4 class \code{\linkS4class{seq_anova_results}}. Click on the
class link to see the full description of the slots.
To get access to the object use the
\code{@}-operator or \verb{[]}-brackets instead of \code{$}.
See the examples below.
}
\description{
Performs a sequential ANOVA.
}
\examples{
# simulate data ----------------------------------------------------------------
# Set coefficients
alpha = 10
# beta1 = .3
beta1 = .0
beta2 = -.5
beta3 = -1.1

# Generate 200 trials
A = c(rep(c(0), 100), rep(c(1), 100), rep(c(2), 100)) # '0' 100 times, '1' 100 times
# B = rep(c(rep(c(0), 50), rep(c(1), 50)), 2) # '0'x50, '1'x50, '0'x50, '1'x50
e = rnorm(300, 0, sd = 1) # Random noise, with standard deviation of 1

# Generate your data using the regression equation
# y = alpha + beta1*A + beta2*B + beta3*A*B + e
y = alpha + beta1*A  + e

# Join the variables in a data frame
# data = data.frame(cbind(A, B, y))
data = data.frame(cbind(A, y))

# Check the data with a regression
# lin.model = lm(y ~ A*B, data = data)
# lin.model = lm(y ~ A, data = data)
# summary(lin.model)


# calculate sequential ANOVA ---------------------------------------------------
results <- sprtt::seq_anova(y ~ A, f = 0.2, data = data )
results@decision
}
