% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_anova.R
\name{seq_anova}
\alias{seq_anova}
\title{Sequential Analysis of Variance}
\usage{
seq_anova(formula, f, alpha = 0.05, power = 0.95, data, verbose = TRUE)
}
\arguments{
\item{formula}{a formula.}

\item{f}{cohens f.}

\item{alpha}{the type I error. A number between 0 and 1.}

\item{power}{1 - beta (beta is the type II error probability). A number
between 0 and 1.}

\item{data}{data frame.}

\item{verbose}{a logical value whether you want a verbose output or not.}
}
\value{
An object of the S4 class \code{\linkS4class{seq_anova_results}}. Click on the
class link to see the full description of the slots.
To get access to the object use the
\code{@}-operator or \verb{[]}-brackets instead of \code{$}.
See the examples below.
}
\description{
Performs a sequential ANOVA.
}
