% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprtt.R
\docType{package}
\name{sprtt}
\alias{sprtt}
\title{sprtt: help page.}
\description{
This package provides the implementation of sequential probability
ratio tests using t-statistic.
}
\section{detailed help pages}{

For detailed instructions on the \code{sprtt} package, see:
\link{https://meikesteinhilber.github.io/sprtt/index.html}
}

\section{sprtt functions}{

\itemize{
\item \code{\link{seq_ttest}} Performs one and two sample sequential t-tests on vectors of
data.
}
}

